requirements.txt
fastmcp>=0.1.0
gradio>=4.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-mcp-adapters>=0.1.0
duckduckgo-search>=6.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
PyYAML>=6.0.0
tiktoken>=0.7.0